Add Gate array, and various clocks. Wait line also.
