+++ AmstradCore/IComponent.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <string>
+#include <vector>
+
+class IComponent
+{
+public:
+   virtual void Tick() = 0;
+};
