+++ AmstradCore/Motherboard.h
@@ -3,57 +3,11 @@
 #include <string>
 #include <vector>
 
-class IComponent
-{
-public:
-   virtual void Tick() = 0;
-};
+#include "IComponent.h"
+#include "BusLine.h"
+#include "GateArray.h"
+#include "SingleLineSample.h"
 
-class BusLine : public IComponent
-{
-public:
-   BusLine();
-   virtual ~BusLine();
-
-   void AddComponent(IComponent* component);
-
-   void Tick() override;
-   bool GetLevel();
-
-protected:
-
-   std::vector<IComponent*> component_list_;
-   bool up_;
-};
-
-class SingleLineSample
-{
-public:
-   SingleLineSample(std::string label, BusLine* line);
-
-   void Clear();
-   std::string GetSample();
-
-   std::string label_;
-   BusLine* line_;
-   std::vector<bool> samples_;
-};
-
-class GateArray : public IComponent
-{
-public:
-   GateArray();
-   virtual ~GateArray();
-
-   void CreateGateArray(BusLine* line_4, BusLine* line_1);
-
-   void Tick() override;
-
-protected:
-   unsigned int counter;
-   BusLine *line_4_mhz_;
-   BusLine *line_1_mhz_;
-};
 
 class Motherboard
 {
@@ -69,6 +23,10 @@ public:
    void Create();
 
    ///////////////////////////////////////
+   // Reset
+   void Reset();
+
+   ///////////////////////////////////////
    // Run
    void Tick();
 
@@ -80,7 +38,7 @@ public:
 protected:
    ///////////////////////////////////////
    // Sample
-   bool sample_;
+   bool sample_{};
 
    std::vector<SingleLineSample> samples_;
 
@@ -88,7 +46,9 @@ protected:
    // Inner components
    BusLine line_16_mhz_;
    BusLine line_4_mhz_;
-   BusLine line_1_mhz_;
+   BusLine line_CCLK_mhz_;
+   BusLine line_wait_;
+
 
    GateArray gate_array_;
 };
